moved a few more files and updated logo
