+++ client_app/src/App.js
@@ -1,7 +1,7 @@
 import { BrowserRouter as Router, Routes, Route } from "react-router-dom";
-import { PastShifts, UpcomingShifts } from "./Shifts";
-import Shelters from "./Shelters";
-import VolunteerDashboard from "./VolunteerDashboard";
+import { PastShifts, UpcomingShifts } from "./components/volunteer/Shifts";
+import Shelters from "./components/volunteer/Shelters";
+import VolunteerDashboard from "./components/volunteer/VolunteerDashboard";
 import NavBarVolunteerDashboard from "./components/volunteer/NavBarVolunteerDashboard";
 import NavBarShelterDashboard from "./components/shelter/NavBarShelterDashboard";
 import Login from "./components/authentication/Login";
@@ -9,9 +9,9 @@ import Logout from "./components/authentication/Logout";
 import SignUp from "./components/authentication/SignUp";
 import ProtectedRoute from "./ProtectedRoute";
 import { useState } from "react";
-import ShelterDashboard from "./ShelterDashboard";
+import ShelterDashboard from "./components/shelter/ShelterDashboard";
 
-import "./App.css";
+import "./styles/App.css";
 
 function App() {
   const [auth, setAuth] = useState(false);
